backend/requirements.txt
django
channels
uvicorn
langchain
langchain-core
langchain-google-genai
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
google-genai